dnd-functionality-typeBugFix
